resolves 84: fixed changes
