+++ frontend/src/services/password-thunk.ts
@@ -1,38 +1,15 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
-
+import { PasswordPayload,
+RotateAESKeyAndIVPayload,
+RotateCharsetThunk,
+EncryptPasswordThunk,
+DecryptPasswordThunk,
+DeletePasswordThunk,
+GenerateBackupThunk,
+DownloadBackupThunk,
+UploadBackupThunk } from "../models/passwordInterface";
 import * as passwordService from "./password-service";
-interface PasswordPayload {
-  identifier: string;
-}
-interface RotateAESKeyAndIVPayload {
-  passUid: string;
-}
-interface RotateCharsetThunk{
-  passUid:string;
 
-}
-interface EncryptPasswordThunk{
-  passUid:string;
-  password:string;
-}
-interface DecryptPasswordThunk{
-  passUid:string;
-  pmsFile:File ;
-
-}
-interface DeletePasswordThunk{
-  passUid:string;
-}
-interface DownloadBackupThunk{
-  backupName:string;
-}
-interface GenerateBackupThunk{
-  password:string ;
-}
-interface UploadBackupThunk{
-  backupFile:File ;
-  password:string ;
-}
 export const createPasswordThunk = createAsyncThunk(
   "password/createPassword",
   async (payload:PasswordPayload) => {
